requirements.txt
flask
flask-cors
python-dotenv
requests
langchain
langchain-cohere
langchain-huggingface
langchain-groq
langchain-chroma
chromadb
gtts
geocoder
transformers
datasets
firebase-admin
pyrebase4
python-jose[cryptography]
Flask-Session
